Add new test EmptyDepAndZeroOutput.
This test trigger a case that was not covered yet by CMake's test suite.
However, CMake's master branch did not have this bug. I caught it while
writing an optimized version of cmMakefile::GetSourceFileWithOutput(). My
first attempt was passing all the test suite but was buggy. I would have
find the bug in the first place if this test was already present. This
patch fills this gap.